+++ ProjectDirectX12/resource.hpp
@@ -10,7 +10,7 @@ namespace pdx12
 	resource_and_state create_commited_upload_buffer_resource(ID3D12Device* device, UINT64 size);
 
 	resource_and_state create_commited_texture_resource(ID3D12Device* device,
-		DXGI_FORMAT format, UINT64 width, UINT64 height, std::size_t dimension, UINT16 depthOrArraySize, UINT16 mipLevels, D3D12_RESOURCE_FLAGS flags, D3D12_CLEAR_VALUE const* clearValue = nullptr);
+		DXGI_FORMAT format, UINT64 width, UINT height, std::size_t dimension, UINT16 depthOrArraySize, UINT16 mipLevels, D3D12_RESOURCE_FLAGS flags, D3D12_CLEAR_VALUE const* clearValue = nullptr);
 
 	resource_and_state create_commited_buffer_resource(ID3D12Device* device, UINT64 size, D3D12_RESOURCE_FLAGS flags);
